Move identifiers.py out of utils directory
